LC-n-th-tribonacci-number [Runtime: 0 ms; Memory: 7.9 MB]

+++ math/n-th-tribonacci-number.cpp
@@ -7,7 +7,7 @@ public:
         else if(n == 2) return 1;
         else if(n == 3) return 2;
 
-        int first = 0, second = 1, third = 1;
+        long long first = 0, second = 1, third = 1;
         long long sum = 0;
 
         for(int i=0; i<n; i++){
